Added actual Python files
Added files. The ANSI art is a random one found in a huge Archive. I am trying to track down the artist. It is over 25 years old. The name 12 Bones of Doom is a TM of Adam Hansen. The code itself is there for you use with your own title name.